+++ TMessagesProj/src/main/java/org/telegram/ui/Components/AudioPlayerAlert.java
@@ -459,8 +459,7 @@ public class AudioPlayerAlert extends BottomSheet implements NotificationCenter.
             } else {
                 TLRPC.Chat chat = MessagesController.getInstance(currentAccount).getChat(-did);
                 if (chat != null) {
-                    String title = UserConfig.getChatTitleOverride(currentAccount, chat.id);
-                    /* MERGE_CONFLICT */
+                    String title = UserConfig.getChatTitleOverride(currentAccount, (int)chat.id);
                     if (title == null) {
                         title = chat.title;
                     }
